+++ client/src/app/App.js
@@ -1,30 +1,24 @@
 import React from 'react';
 import {Route, Routes} from 'react-router-dom';
 import {ToastContainer} from 'react-toastify';
-import useAuth from './util/AuthContext';
-import NavBarLoggedIn from './components/NavBarLoggedIn';
-import NavBarLoggedOut from './components/NavBarLoggedOut';
+import Box from '@mui/material/Box';
 import Landing from './pages/Landing';
 import MyProfile from './pages/MyProfile';
-import Opportunities from './pages/Opportunities';
-import Dashboard from './pages/Dashboard';
-import Approvals from './pages/Approvals';
-// import Opportunity from './pages/Opportunity';
-import Profile from './components/Profile';
-import 'react-toastify/dist/ReactToastify.css';
-import './stylesheets/App.css';
-
 import Signup from './pages/Signup';
 import Login from './pages/Login';
 import Verify from './components/Verify';
-
-// TODO: delete browse page
-import Browse from './pages/Browse';
-// TODO: settings page?
+import NavBarLoggedOut from './components/NavBarLoggedOut';
+import NavBarLoggedIn from './components/NavBarLoggedIn';
+import Dashboard from './pages/Dashboard';
+import Approvals from './pages/Approvals';
+import Opportunities from './pages/Opportunities';
 import Settings from './pages/Settings';
-import Box from '@mui/material/Box';
-
 import ViewOpportunity from './pages/ViewOpportunity';
+import Profile from './components/Profile';
+import Browse from './pages/_OLD_Browse';
+import useAuth from './util/AuthContext';
+import 'react-toastify/dist/ReactToastify.css';
+import './stylesheets/App.css';
 
 /**
  * returns basic routes and navbar of app
@@ -41,26 +35,24 @@ export default function App() {
         <Routes>
           <Route path='/' element={<Landing />}/>
           <Route path='/myprofile' element={<MyProfile />} />
+
+          <Route path='/signup' element={<Signup />} />
+          <Route path='/login' element={<Login />} />
+          <Route path='/verify/:token' element={<Verify />} />
+
           <Route path='/dashboard' element={<Dashboard/>}/>
           <Route path='/approvals' element={<Approvals/>}/>
           <Route path='/opportunities' element={<Opportunities/>}/>
-          {/*
-            <Route path='/opportunity/:opportunityid' element={<Opportunity/>}/>
-          */}
+          <Route path='/settings' element={<Settings />}/>
+
           <Route
             path='/opportunity/:opportunityid'
             element={<ViewOpportunity />}
           />
           <Route path='/profile/:profileid' element={<Profile />} />
-          {/* TODO: delete browse page */}
-          <Route path='/browse' element={<Browse />}/>
-          {/* TODO: settings page? */}
-          <Route path='/settings' element={<Settings />}/>
 
-          <Route path='/opportunities' element={<Opportunities />}/>
-          <Route path='/signup' element={<Signup />} />
-          <Route path='/login' element={<Login />} />
-          <Route path='/verify/:token' element={<Verify />} />
+          {/* Reimplemented - This is still here for reference */}
+          <Route path='/browse' element={<Browse />}/>
         </Routes>
       </Box>
     </Box>
