+++ client/src/app/components/_OLD_OpportunityDetails.js
