+++ client/src/app/components/_OLD_OpportunityRequests.js
@@ -1,9 +1,10 @@
 import * as React from 'react';
 import {useState, useEffect} from 'react';
 import {Grid} from '@mui/material';
-import OpportunityRequestsCreatorView from './OpportunityRequestsCreatorView';
+import OpportunityRequestsCreatorView
+  from './_OLD_OpportunityRequestsCreatorView';
 import OpportunityRequestsVolunteerView
-  from './OpportunityRequestsVolunteerView';
+  from './_OLD_OpportunityRequestsVolunteerView';
 import useAuth from '../util/AuthContext';
 
 /**
