+++ client/src/app/components/_OLD_PendingOpportunities.js
@@ -1,30 +1,28 @@
 import * as React from 'react';
-import {List} from '@mui/material';
 import Paper from '@mui/material/Paper';
-import OpportunityCard from './OpportunityCard';
-import '../stylesheets/Opportunities.css';
+import '../stylesheets/_OLD_Opportunities.css';
+import {List} from '@mui/material';
+import OpportunityCard from './_OLD_OpportunityCard';
 
 /**
  * creates Profile
  * @return {HTML} Profile component
  */
-export default function JoinedOpportunities({data}) {
-  // console.log(data);
-
+export default function PendingOpportunities({data}) {
   return (
     <Paper
-      className='joined-opportunities'
+      className='pending-opportunities'
       elevation={3}
       sx={{
-        marginBlock: '3rem',
+        marginBottom: '3rem',
         width: '850px',
         height: 'auto',
         boxShadow: '0px 0px 50px -14px rgba(0, 0, 0, 0.1)',
         borderRadius: '10px',
       }}
     >
-      <div className='joined card-title'>
-        Joined Opportunities
+      <div className='pending card-title'>
+        Pending Opportunities
       </div>
 
       <List
@@ -38,12 +36,13 @@ export default function JoinedOpportunities({data}) {
         }}
       >
         {data && data.map((opportunity, index) => (
-          <OpportunityCard
-            key={`opportunity-list-item-${index}`}
-            data={opportunity}
-          />
+          <OpportunityCard data={opportunity} key={index} />
         ))}
       </List>
+
+      {(!data || data.length == 0) && <h2 className='no_results_message'>
+        No Pending Opportunities found
+      </h2>}
     </Paper>
   );
 }
