+++ client/src/app/components/_OLD_TabBar.js
