--- client/src/app/pages/GetStarted.js
@@ -1,14 +1,0 @@
-import * as React from 'react';
-import '../stylesheets/GetStarted.css';
-
-/**
- * creates get started page
- * @return {HTML} get started page
- */
-export default function GetStarted() {
-  return (
-    <div className='GetStarted'>
-      <h1>Get Started</h1>
-    </div>
-  );
-}
