--- client/src/app/stylesheets/GetStarted.css
@@ -1,3 +1,0 @@
-.GetStarted {
-    padding-top: 75px;
-}
