+++ client/src/app/stylesheets/_OLD_OpportunityCard.css
